Add AI, automation, enterprise, sync, remote, marketplace, reporting, overlay, and monitoring features
- AI anomaly detection, optimizer, tuner, smart alerts
- Automation engine with conditions, actions, events, scheduler
- Enterprise features with RBAC and license management
- Remote server with JWT auth
- Sync client/server for multi-device
- Marketplace for plugin sharing
- Reporting engine with scheduling
- Overlay controller with game detection
- Hardware monitoring with NVIDIA and simulated backends
- Backup engine with rotation
- Integration config system
- Frontend pages for all new features
- Database tests

+++ apps/backend/src/automation/condition.rs
@@ -1,4 +1,3 @@
-use crate::hardware::GpuSample;
 use crate::monitoring::MonitoringEngine;
 use super::rule::{Comparison, Condition};
 
@@ -33,13 +32,13 @@ impl ConditionEvaluator {
     fn extract_metric(metric: &str, sample: &crate::hardware::GpuSample) -> Option<f64> {
         match metric {
             "temperature" | "temp" => Some(sample.temperature_celsius),
-            "core_clock" | "core_mhz" => Some(sample.core_clock_mhz as f64),
-            "memory_clock" | "mem_mhz" => Some(sample.memory_clock_mhz as f64),
-            "fan_speed" | "fan" | "fan_percent" => Some(sample.fan_speed_percent as f64),
+            "core_clock" | "core_mhz" => Some(sample.core_clock_mhz),
+            "memory_clock" | "mem_mhz" => Some(sample.memory_clock_mhz),
+            "fan_speed" | "fan" | "fan_percent" => Some(sample.fan_speed_percent),
             "power" | "power_watts" => Some(sample.power_watts),
             "core_util" | "gpu_util" | "utilization" => Some(sample.core_utilization_percent),
             "mem_util" | "memory_util" => Some(sample.memory_utilization_percent),
-            "core_voltage" | "voltage" => Some(sample.core_voltage_mv as f64),
+            "core_voltage" | "voltage" => Some(sample.core_voltage_mv),
             "mem_used" | "memory_used_mb" => Some(sample.memory_used_mb as f64),
             "mem_total" | "memory_total_mb" => Some(sample.memory_total_mb as f64),
             _ => None,
